Compiler fixes

+++ src/music.h
@@ -3,8 +3,8 @@
 
 #include "types.h"
 
-void sound_play1(u8 sound);
-void sound_play2(u8 sound);
+void sound_play1(u16 sound);
+void sound_play2(u16 sound);
 void update_audio(void);
 
 #endif /* MUSIC_H */
